// packages/philjs-3d/src/webgl/types.d.ts
export interface Camera {
    position: Float32Array;
    target: Float32Array;
    up: Float32Array;
    fov: number;
    aspect: number;
    near: number;
    far: number;
    viewMatrix: Float32Array;
    projectionMatrix: Float32Array;
}
export interface CameraOptions {
    position?: [number, number, number];
    target?: [number, number, number];
    up?: [number, number, number];
    fov?: number;
    aspect?: number;
    near?: number;
    far?: number;
}